requirements.txt
pyperclip>=1.8.1
dearpygui>=0.6.32
PyYAML>=5.3.1